+++ CHANGELOG.md
@@ -5,7 +5,12 @@
   
   Author: **Vladimir Mandic <mandic00@live.com>**  
   License: **MIT**  
-  Repository: **<localhost>**  
+  Repository: **<https://github.com/vladmandic/segmentation.git>**  
   
 ## Changelog
   
+### **HEAD -> main** 2022/09/29 mandic00@live.com
+
+- upload model
+- working prototype
+- initial commit